Drop support for the Kinect camera and add support for the Asus Xtion Pro.
Large amounts of cleanup needed, lots of old API that will crash if called (no
hardware to send the message to!), but glview displays just the depth image
with the Xtion camera now.

This patch is not intended to be merged upstream, but to show that libfreenect
works with the Asus camera with few modifications.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>